fall back to dummy values in ci
