+++ solana-programs/anchor/audius-data/lib/constants.ts
@@ -1,0 +1,5 @@
+export const DUMMY_CN_WALLET_ADDRESSES = [
+    "0x25A3Acd4758Ab107ea0Bd739382B8130cD1F204d",
+    "0x71B55d7bDe40D751087A27e2072F0fF8cacA400a",
+    "0xb4bD6911d3F633A1F7B14D955E68061F6f845027"
+]
